refactor: move default theme classes into defaults package
