+++ common/src/main/java/com/klikli_dev/modonomicon/book/Book.java
@@ -13,7 +13,7 @@ import com.klikli_dev.modonomicon.book.error.BookErrorManager;
 import com.klikli_dev.modonomicon.client.gui.book.BookAddress;
 import com.klikli_dev.modonomicon.client.gui.book.markdown.BookTextRenderer;
 import com.klikli_dev.modonomicon.client.gui.book.theme.BookTheme;
-import com.klikli_dev.modonomicon.client.gui.book.theme.DefaultBookTheme;
+import com.klikli_dev.modonomicon.client.gui.book.theme.defaults.DefaultBookTheme;
 import com.klikli_dev.modonomicon.data.BookEntryJsonLoader;
 import com.klikli_dev.modonomicon.util.BookGsonHelper;
 import it.unimi.dsi.fastutil.objects.Object2ObjectMaps;
