+++ common/src/main/java/com/klikli_dev/modonomicon/client/gui/book/theme/defaults/DefaultBookSkinTheme.java
@@ -4,8 +4,19 @@
  * SPDX-License-Identifier: MIT
  */
 
-package com.klikli_dev.modonomicon.client.gui.book.theme;
+package com.klikli_dev.modonomicon.client.gui.book.theme.defaults;
 
+import com.klikli_dev.modonomicon.client.gui.book.theme.BookContentTheme;
+import com.klikli_dev.modonomicon.client.gui.book.theme.BookFrameTheme;
+import com.klikli_dev.modonomicon.client.gui.book.theme.BookNodeTheme;
+import com.klikli_dev.modonomicon.client.gui.book.theme.BookOverviewTheme;
+import com.klikli_dev.modonomicon.client.gui.book.theme.BookRecipeTheme;
+import com.klikli_dev.modonomicon.client.gui.book.theme.BookSkinTheme;
+import com.klikli_dev.modonomicon.client.gui.book.theme.GuiButtonSprites;
+import com.klikli_dev.modonomicon.client.gui.book.theme.GuiFrameOverlay;
+import com.klikli_dev.modonomicon.client.gui.book.theme.GuiNineSlice;
+import com.klikli_dev.modonomicon.client.gui.book.theme.GuiSprite;
+import com.klikli_dev.modonomicon.client.gui.book.theme.GuiTexture;
 import net.minecraft.resources.Identifier;
 
 public final class DefaultBookSkinTheme implements BookSkinTheme {
